+++ src/solvers/linear_system_solver.h
@@ -8,6 +8,6 @@
 #include "../grid/grid/grid.h"
 #include <stdbool.h>
 
-uint64_t conjugate_gradient(struct grid* the_grid, int max_its, double tol, bool use_jacobi, double *error);
+uint32_t conjugate_gradient(struct grid *the_grid, int max_its, double tol, bool use_jacobi, double *error);
 
 #endif //MONOALG3D_LINEAR_SYSTEM_SOLVER_C_H
